requirements.txt
python-vlc
youtube-dl
pymongo
dnspython